separate out ce template validations into ce specific validator
